backend/requirements.txt
# Dependencias para SAMABOT PLC Backend
# Backend REST con FastAPI y comunicación PLC Siemens

# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Comunicación PLC
python-snap7==1.3

# Validación de datos
pydantic==2.5.0

# Logging y utilidades
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Desarrollo
black==23.11.0
flake8==6.1.0
mypy==1.7.1 